+++ NEWS.md
@@ -1,3 +1,10 @@
+# rolap 2.3.0
+* Change star database list as parameter in `constellation()` function.
+* Fix `replace_empty_values()` with spaces at the beginning or end of non-empty strings.
+* Star databases data incremental refresh functionality: `update_according_to()`, `get_transformation_code()`, `get_new_dimension_instances()` and `incremental_refresh()` among other functions.
+* Vignette about star databases incremental refresh.
+* Fix new problem building macos-latest (release) with input functions from file or database in vignettes.
+
 # rolap 2.2.0
 * Vignette about obtaining and transforming flat tables.
 * Definition of flat_table class with functions to transform an integrate data.